Hadar/Shira/Shifi/login-and-signup-in-server

+++ server/EnglishGamesPlatform.Backend/Controllers/GameControllerBase.cs
@@ -23,7 +23,7 @@ namespace EnglishGamesPlatform.Backend.Controllers
             if (response.IsSuccess)
                 return Ok(response);
             else
-                return StatusCode(response.StatusCode, response);
+                return StatusCode((int)response.StatusCode, response);
         }
 
         [HttpGet("leaderboard")]
@@ -33,7 +33,7 @@ namespace EnglishGamesPlatform.Backend.Controllers
             if (response.IsSuccess)
                 return Ok(response);
             else
-                return StatusCode(response.StatusCode, response);
+                return StatusCode((int)response.StatusCode, response);
         }
     }
 }
